Share ioctl code between ATA CF and BIOSHD drivers
